API specification
This commit implements the API specification.
The specification can be read easily on
http://kairo.eti.br/ownca-rest-api/ or running the local HTTP
API locally, after merge it will be available on
http://github.io/ownca-rest-api/ as well.
```
pipenv shell
pipenv install
flask run
```

This commit has no functional implementation, and the idea is to
review the API specification until it gets ready and then create
the issues related to the implementation.

Signed-off-by: Kairo de Araujo <kairo@dearaujo.nl>